Velocity and Degrees Work

+++ TransponderReceiverUser/Plane/IPlane.cs
@@ -9,8 +9,8 @@ namespace TransponderReceiverUser
         int XPos { get; set; }
         int YPos { get; set; }
         int Altitude { get; set; }
-        int Velocity { get; set; }
-        int Degrees { get; set; }
+        double Velocity { get; set; }
+        double Degrees { get; set; }
         bool ConditionCheck { get; set; }
         DateTime TimeStamp { get; set; }
     }
